pyproject.toml
[tool.poetry]
name = "todolist"
version = "0.1.0"
description = "todolist start les33"
authors = ["MrLerich <mrlerich@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.5"
envparse = "^0.2.0"
psycopg2-binary = "^2.9.5"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.13.0"
social-auth-app-django = "^5.0.0"
djangorestframework-jsonapi = "^6.0.0"
django-filter = "^22.1"
django-social-oauth-token = "^2.2.0"
marshmallow = "^3.19.0"
marshmallow-dataclass = "^8.5.11"


[tool.poetry.group.dev.dependencies]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
